Add files via upload
[Wass2,WassSep,SinkSep,Mean2Sep]=bwsi(x,y,M,sigma).
This subroutine computes the Wasserstein-Bures Estimators. You need to input the x matrix of your inputs and the y matrix of your output. The the subroutines selects the number of partitions. However, you may also set the number M as a third argument Sigma stands for the parameter of the entropic approximation with the Bures-Sinkhorn entropy instead of the Wasserstein.
The outputs are Wass2 (the Wasserstein 2 distance, the squares of the sepatations (WassSep) and the same for the approximate Sinkhorn-entropy estimator; Mean2Sep gives the advective part).